fix compile error related to ssize_t.

+++ extensions/GUI/CCScrollView/CCTableView.cpp
@@ -51,7 +51,7 @@ bool TableView::initWithViewSize(Size size, Node* container/* = NULL*/)
 {
     if (ScrollView::initWithViewSize(size,container))
     {
-        _indices        = new std::set<long>();
+        _indices        = new std::set<ssize_t>();
         _vordering      = VerticalFillOrder::BOTTOM_UP;
         this->setDirection(Direction::VERTICAL);
